refactor: reorganize clientHelpers and move cosineSimilarity to stringHelpers
- Removed cosineSimilarity and splitMessage functions from index.ts and relocated them to a new stringHelpers.ts file for better modularity.
- Updated index.ts to export from stringHelpers.ts and MsgPreprocessor.ts, improving code organization and clarity.
